pyproject.toml
[project]
name = "fatebook-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "mcp[cli]>=1.12.4",
    "python-dotenv>=1.1.1",
]
